fix matching file logic
There is a case where the file is in multiple collection, one the current upload collection (say x ) and one another collection (say y).

we might get y collection's file and we will try to add a symlink for it in collection x and that will return a no-op from the server , and we would show the file as uploaded, instead of showing it as ignored already in collection.