Remove out of date user contributed ports
As nice it is to have user contributed ports, the problem is that they're
getting outdated. This one is already a few versions behind our latest release,
and we don't have way to update the port so that it's up to date.

Thus removing it from the README so that we're not accidentally suggesting
people to install outdated versions with bugs and/or missing features.