Use a template task bar icon on macOS
On macOS, Electron supports providing a "template" icon that the system will
then automatically color appropriately depending on the light / dark theme.

> Template images consist of black and an alpha channel. Template images are not
  intended to be used as standalone images and are usually mixed with other
  content to create the desired final appearance.
>
>
> The most common case is to use template images for a menu bar icon, so it can
  adapt to both light and dark menu bars.
>
> Note: Template image is only supported on macOS.
>
> To mark an image as a template image, its filename should end with the word Template.

Maybe the same icon can also be used in other platforms, but I'm not aware of
why the existing icon is white, so I've added a separate file so that the rest
of the platforms remain untouched, and only macOS gets this light / dark mode
fix.

References:

* https://www.electronjs.org/docs/latest/api/tray
* https://www.electronjs.org/docs/latest/api/native-image#template-image