Add the TCPShield logo (#408)
<!--
  Thanks for contributing!

Provide a description of your changes below and a general summary in the
title

Please look at the following checklist to ensure that your PR can be
accepted quickly:
-->

## Description

<!--- Describe your changes in detail -->

This PR adds the TCPShield. TCPShield is a highly available DDoS
protection platform, mainly used in Minecraft servers. I have also
re-aranged NextDNS and Skiff in the custom-icons.json to make it in
alphabetical order.

<!--- Put an `x` in all the boxes that apply: -->

- [x] 🖼️ New icon
- [ ] ✨ New feature (non-breaking change which adds functionality)
- [ ] 🛠️ Bug fix (non-breaking change which fixes an issue)
- [ ] ❌ Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] 🧹 Code refactor
- [ ] ✅ Build configuration change
- [ ] 📝 Documentation
- [ ] 🗑️ Chore