Update the UI when deleting a subset of the duplicates
The list of photos shown on the dedup screen did not update if we selected a
subset of the duplicates and deleted them. This is because while we're issuing a
remote API request, we don't seem to be updating our local storage with the
results of that when we're on this screen. So when the gallery on this screen
re-renders, it still reads the same files again, even after they've been
deleted.

Not happy with the current solution, but it does solve the customers issue. Have
added an "Enhancement" tag to flag this.

Test:

I was able to reproduce the issue. After making these changes, I don't see it
anymore. Haven't done exhaustive checks, but it fixes the happy path at least.

Tested locally.