Upgrade next (remaining in 13 series) to fix moduleResolution issue
This is to prevent Next from forcefully overwriting the "moduleResolution" in
tsconfig.json to "node". If it is "node", then VS code cannot find the imports
we do from @repo/ui.

This was a bug in Next which was fixed in this PR:
https://github.com/vercel/next.js/pull/51957

To upgrade, I used:

    yarn upgrade next@^13.5

The latest Next.js is 14, but I left it at the last release in the 13
series. Will deal with the upgrade to 14 later in a different PR.