Fix `yarn workspace photos add foo`
This was likely happening because the package versions in the top level
package.json and the eslint-config package.json were different. I've synced them
for now.

> On a longer time frame, the idea is not have such conflicting cases, and
remove mostly all the top level package.json dependencies and for the rest rely
on them solely being in the top level package.json and not duplicated in two
places.