Remove the need for nested child selectors to workaround emotional (sic) issues
This workaround was added in https://github.com/ente-io/photos-web/pull/1027,
but it has stopped working again, likely due to the update to Next 14. We're
only using nested child selectors in one place (not 100% sure of this, but I did
try to poke around and verify that statement a bit).

This was causing an error in dev mode, and silent failure in production mode.

- Hover state is restored
- dev mode seems to work fine
- yarn build also works