Allow developers to customize their VSCode configuration
For folks who're using the git integration from within VSCode, the top level
directory needs to be open in VSCode for using git GUI from VSCode.

However, for using the Run & debug configurations they need to open ente/mobile
is opened in VSCode.

As a solution

- gitignore .vscode
- provide sample VSCode configuration files is mobile/docs/vscode and auth/docs/vscode
- mention it in the README

(will update the auth README later, have to do a few more changes to it first)