feat(server): Add ability to configure sslmode
This patch makes the sslmode for the postgres database configurable. It
adds a new configuration option to for the database setup and removes
the hardcoded value. This change is done to allow usage of
hosted-databases that require tls for connections, like ones created by
the zalando postgres operator.

Alternatively one could use the Postgres standardised environment
variables, but this would probably break the current behaviour.

References:
https://postgres-operator.readthedocs.io/en/latest/user/#connect-to-postgresql
https://www.postgresql.org/docs/current/libpq-envars.html#LIBPQ-ENVARS