[workflow] Retain the PlayStore AAB as a workflow artifact
Reasoning in the workflow comment.

While this wouldn't have helped in cases like the recent workflow failure
(https://github.com/ente-io/ente/actions/runs/8211812037/job/22461255230), it is
maybe a good idea to attach the AAB to the workflow artifacts so that we can get
at the binary if needed.

I'm not fully sure if we need this though. So we can close this PR too - I just
wanted to figure how to do this if we need to make this amendment later.

Untested. I validated the YAML though.

    yq eval .github/workflows/auth-release.yml