Specify which Google Takeout format is supported
Currently, the upload menu says it accepts Google Takeout.
Unfortunately, it's not clear to the user which format, ZIP or TGZ, is
actually supported. This can lead a user to accidentally request Takeout
in the wrong format, which is especially painful if the user has a large
library that takes a long time to export.

Perhaps both ZIP and TGZ should be supported. At the very least, this
change explicitly tells the user which format is supported, so they
don't waste a day exporting the wrong format.