[web] Use the public link origin provided by the server
When implementing https://github.com/ente-io/ente/pull/1443 it was found that
the Photos web app wasnt' using the sharedCollectionUrl prefix that is returned
by the server. Instead, it was using NEXT_PUBLIC_ENTE_ALBUMS_ENDPOINT as the
baseHost, while mobile client completely rely on the server for returning the
correct hostUrl.

**Tested by**

Running against the modified server, and verifying that changes to
apps.public-albums were being used.