[web] Allow thumbnail cache to be optional
See: https://github.com/ente-io/ente/discussions/1449#discussioncomment-9255346

I'm not yet sure what was the case why it was not initialized, hoping to get
some logs for the error when initializing the cache to see how we got to this
state. But meanwhile, ensure that the code works even without the cache.