Rework the video chunk decryptor stream logic
When running on Ubuntu 24 arm64 in the desktop app (didn't test on web0, trying
to open certain videos fails with:

> [rndr] [error] Failed to process file stream: TypeError: Failed to execute
  'enqueue' on 'ReadableStreamDefaultController': Cannot enqueue a chunk into a
  closed readable stream

While not specifically fixing that issue, I'm first rewriting this to use the
more normal (recommended?) approach of implementing a pull instead of doing
everything in start. Maybe that fixes the issue, otherwise at least one less
ghost for me to worry about.