[web][photos] Fix mobile upload button hover & position
The hover and ripple effects and the positioning were off on the mobile upload button, because the `.mobile-button` class used the wrong class to let the element appear again. It used `display: block`, but the MUI IconButton actually has `display: inline-flex`, which will also fix the wrong height of the button.

See https://github.com/mui/material-ui/issues/33020