[desktop] Fix broken nightly builds due to ajv
I'm not sure why it started happening now. Earlier I thought it was because of
this - https://github.com/ente-io/ente/pull/2969 - but that was a red-herring (I
think!), instead this likely got triggered because of
https://github.com/ente-io/action-electron-builder/commit/eff78a1d33bdab4c54ede0e5cdc71e0c2cf803e2.

This change fixed the issue locally, will see if it works on CI too.

Workaround from:
https://github.com/ajv-validator/ajv/issues/2443#issuecomment-2147026958