[cli] sync deleted files before syncing new/updates files (#4776)
## Description
The CLI sorts deleted photos to the bottom of the sync queue. 
When an album is synced, a photo is removed from the album and later
re-added to the album (same name) and doing another sync the new file is
added first and a new name is generated since the original filename is
(still) taken. Afterwards the initial photo is deleted and the filename
would be available.
This leads to having a file, e.g. IMG_0001_1.JPG instead of the original
filename IMG_0001.JPG despite the initial filename is no longer on disk
and could have been used.

This PR changes the sort order so that deleted files are first removed
and afterwards new files are created. In this case all files where the
filename is now available, but were taken, are named like they were
uploaded.