[quickstart] Gracefully handle case when docker compose is not present
When docker is present but docker compose is not present, the `docker compose`
invocation would fail. We want the early exit (`set -e`), so instead do a
fallback to set dcv to an empty string so that it later fails in the `test -z
dcv` case below and prints the intended error message.